feat: update values
